Engine modifications and docker cluster updates

+++ streamflow-core/streamflow-model/src/main/java/streamflow/model/ComponentPropertyOptions.java
@@ -33,7 +33,7 @@ public class ComponentPropertyOptions implements Serializable {
 
     private Integer minNumber;
 
-    private Float floatStep;
+    private Double floatStep;
 
     private String displayUnits;
 
@@ -77,11 +77,11 @@ public class ComponentPropertyOptions implements Serializable {
         this.minNumber = minNumber;
     }
 
-    public Float getFloatStep() {
+    public Double getFloatStep() {
         return floatStep;
     }
 
-    public void setFloatStep(Float floatStep) {
+    public void setFloatStep(Double floatStep) {
         this.floatStep = floatStep;
     }
